[PR] Add cell_dci_t message type
Send the cell_dci_t message instead of ue_dci. cell_dci_t includes the
information of all RNTIs in a DCI such as dl/dl TBS, PRBs and the
number of not-decoded PRBs (where the TBS is 0, but PRBs are used).

PR: https://github.com/YaxiongXiePrinceton/NG-Scope/pull/14/files